requirements.txt
biopython==1.84
certifi==2024.12.14
charset-normalizer==3.4.0
idna==3.10
numpy==2.2.0
requests==2.32.3
urllib3==2.2.3
